Connector modules restructurized
