+++ Procfile
@@ -1,2 +1,2 @@
-web: NODE_ENV=production node --optimize_for_size --max_old_space_size=$MEMORY_AVAILABLE -r newrelic build/web
-worker: NODE_ENV=production node --optimize_for_size --max_old_space_size=$MEMORY_AVAILABLE -r newrelic build/worker
+web: NODE_ENV=production WEB=true node --optimize_for_size --max_old_space_size=$MEMORY_AVAILABLE -r newrelic build
+worker: NODE_ENV=production WORKER=true node --optimize_for_size --max_old_space_size=$MEMORY_AVAILABLE -r newrelic build
