+++ package.json
@@ -7,9 +7,9 @@
     "build": "npm run clean && npm run build:server",
     "build:server": "babel server -d build",
     "start": "npm run build && COMBINED=true node build/web",
-    "start:dev": "NODE_ENV=development COMBINED=true babel-watch -L server/web",
-    "test": "npm run test:modules && npm run test:lint && npm run test:flow",
-    "test:modules": "mocha --require babel-register -R spec ./tests",
+    "start:dev": "NODE_ENV=development COMBINED=true babel-watch -L server",
+    "test": " npm run test:lint && npm run test:flow && npm run test:units",
+    "test:units": "mocha --require babel-register -R spec ./tests",
     "test:lint": "eslint server tests",
     "test:flow": "flow check",
     "clean": "rm -rf build",
@@ -38,23 +38,20 @@
     "JSONStream": "^1.1.1",
     "babel-cli": "^6.14.0",
     "babel-preset-es2015": "^6.3.13",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-stage-0": "^6.3.13",
     "babel-register": "^6.4.3",
     "basic-auth": "^1.0.4",
     "batch-stream": "^0.1.3",
     "bluebird": "^3.3.5",
     "body-parser": "^1.15.1",
-    "cache-manager": "^2.2.0",
     "connect-timeout": "^1.8.0",
     "cors": "^2.8.1",
     "csv-stream": "^0.1.3",
     "datadog-metrics": "^0.3.0",
     "ejs": "^2.4.1",
     "express": "^4.13.4",
-    "hull": "^0.11.0-beta.4",
-    "kue": "^0.11.1",
-    "kue-ui": "^0.1.0",
-    "librato-node": "^4.1.0",
+    "hull": "^0.11.0",
     "lodash": "^4.12.0",
     "moment": "^2.14.1",
     "newrelic": "^1.28.3",
@@ -65,16 +62,15 @@
     "superagent": "^2.2.0",
     "superagent-prefix": "0.0.2",
     "superagent-promise-plugin": "^3.2.0",
-    "supply": "0.0.4",
     "urijs": "^1.18.0"
   },
   "devDependencies": {
     "babel-eslint": "^6.0.4",
-    "babel-preset-flow": "^6.23.0",
     "babel-watch": "^2.0.2",
     "eslint": "^2.9.0",
     "eslint-config-airbnb-base": "^3.0.1",
     "eslint-plugin-import": "^1.8.0",
+    "flow-bin": "^0.46.0",
     "mocha": "^2.5.2",
     "nock": "^9.0.13",
     "sinon": "^1.17.4"
