+++ server/actions/batch-handler.js
@@ -1,0 +1,6 @@
+export default function batchHandler(req, res, next) {
+  const segmentId = req.query.segment_id || null;
+  const source = req.query.source || null;
+  req.hull.enqueue("handleBatch", { body: req.body, segmentId, source })
+    .then(next, next);
+}
