+++ server/actions/fetch-all.js
@@ -1,0 +1,4 @@
+export default function fetchAll(req, res, next) {
+  req.hull.enqueue("fetchAllUsers")
+    .then(next, next);
+}
