+++ server/actions/sync.js
@@ -1,0 +1,4 @@
+export default function sync(req, res, next) {
+  return req.hull.enqueue("fetchUsers")
+    .then(next, next);
+}
