--- server/controller/index.js
@@ -1,3 +1,0 @@
-export Actions from "./actions";
-export Jobs from "./jobs";
-export NotifHandlers from "./notif-handlers";
