+++ server/notif-handlers/index.js
@@ -1,0 +1,7 @@
+export shipUpdate from "./ship-update";
+
+export userUpdate from "./user-update";
+
+export segmentUpdate from "./segment-update";
+export segmentDelete from "./segment-delete";
+
