+++ server/notif-handlers/segment-delete.js
@@ -1,0 +1,11 @@
+import Promise from "bluebird";
+
+export default function segmentDelete(ctx) {
+  const { syncAgent } = ctx.shipApp;
+  if (!syncAgent.isConfigured()) {
+    ctx.client.logger.info("ship is not configured");
+    return Promise.resolve();
+  }
+
+  return syncAgent.syncShip();
+}
